Create template - Table of contents
When you use this template it will take a look at the current directory any folders inside that directory will become headers to the content table it will then link every file in that directory for you. After which it stamps the time you had last updated this. keep in mind it will sort them alphabetically so if you have calculus I, calculus II, calculus III it will present in the order of the courses.